Sharing Options from Backend
