+++ src/shared/fake/fake-data.ts
@@ -1,8 +1,9 @@
 import { Observable } from 'rxjs';
 import { CUImage, CUAlbum, CUUser, CUFile } from './../models/_cu-models.provider';
-import { CUUserPreferences, CUSharingOption } from './../../shared/models/cu-user-preferences.class';
-import { QuickShare } from '../../shared/services/share/quick-share.option';
-import { FullShare } from '../../shared/services/share/full-share.option';
+import { SharingOption } from './../models/sharing/_sharing-options';
+import { CUUserPreferences } from './../../shared/models/cu-user-preferences.class';
+import { QuickShare } from '../../shared/models/sharing/quick-share.option';
+import { FullShare } from '../../shared/models/sharing/full-share.option';
 
 export class FakeData {
 
@@ -193,19 +194,22 @@ export class FakeData {
         ];
     }
 
-    public static get Preferences(): CUUserPreferences {
-        let prefs: CUUserPreferences = {
+    public static get Preferences(): {} {
+        let prefs = {
             sharingOptions: [
                 {
-                    name: 'Direct',
-                    description: 'A link directly to the image',
-                    strategy: new QuickShare(),
-                    order: 0
-                }, {
                     name: 'Share',
                     description: 'A link to the image in a page with the title and description',
-                    strategy: new FullShare(),
-                    order: 1
+                    enabled: true,
+                    order: 0,
+                    strategyId: '1002',
+                },
+                {
+                    name: 'Share',
+                    description: 'A link to the image in a page with the title and description',
+                    enabled: true,
+                    order: 0,
+                    strategyId: '1001',
                 }
             ]
         }
