--- src/shared/models/sharing/cu-sharing-option.class.ts
@@ -1,17 +1,0 @@
-import { Constants } from './../../constants';
-
-export class SharingOption {
-    name: string;
-    description: string;
-    enabled: boolean;
-    order: number;
-    link: string;
-
-    constructor(name, description, enabled, order, url){
-        this.name = name;
-        this.description = description;
-        this.enabled = enabled;
-        this.order = order;
-        this.link = 'll';
-    }
-}
