+++ src/shared/models/sharing/quick-share.option.ts
@@ -1,0 +1,9 @@
+import { SharingOptionInterface } from './_sharing-options';
+
+export class QuickShare implements SharingOptionInterface {
+    id: string = '1001';
+
+    public getLink(url: string): string {
+        return url + ' lol quick share';
+    }
+}
